+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # chat21-ionic ver 1.0
 
 ### 0.934
+- link projectId open in _blank
+- change the color of the text message to in the selected conversation
+- fixes bug on metadata conversation
+- fixes bug multiple sound on changed / added message
+
+### 0.933
 - add user status in header detail conversation
 - modified css detail conversation
 - adds CHANGELOG.md